RC4 1.0.0
At this point it can encrypt following the rc4 algorithm, with its KSA and its PRGA. I SPENT WAY TOO MUCH TIME TRYING TO FIX IT WHILE IT WAS ALREADY WORKING, i just realized i was appending to the keystream the element of S[i] WHEN IT WAS SUPPOSED TO BE S[t] (in the prga). It's ok now, it works.
I'll be adding a menu one day.
Next goal: decrypt.